Minor change on function delete

+++ plugins/main_sections/ms_plugins/functions_delete.php
@@ -40,8 +40,11 @@ function delete_plugin($pluginid){
 	
 	rrmdir(MAIN_SECTIONS_DIR."ms_".$anwser['name']);
 	rrmdir(PLUGINS_DIR."computer_detail/cd_".$anwser['name']);
-	unlink(PLUGINS_SRV_SIDE.$anwser['name'].".zip");
-		
+	
+	if(file_exists(PLUGINS_SRV_SIDE.$anwser['name'].".zip")){
+		unlink(PLUGINS_SRV_SIDE.$anwser['name'].".zip");
+	}
+	
 	$conn->query("DELETE FROM `".DB_NAME."`.`plugins` WHERE `plugins`.`id` = ".$pluginid." ");
 	
 }
